cleanup dependencies and move some code around.
