+++ pg_search/Cargo.toml
@@ -20,18 +20,20 @@ icu = ["tokenizers/icu"]
 
 [dependencies]
 anyhow = { version = "1.0.87", features = ["backtrace"] }
+chrono = "0.4.38"
 crossbeam = "0.8.4"
-csv = "1.3.0"
 derive_more = "0.99.18"
 fs2 = "0.4.3"
-indexmap = "2.5.0"
+humansize = "2.1.3"
 json5 = "0.4.1"
 libc = "0.2.158"
 memoffset = "0.9.1"
 once_cell = "1.19.0"
+os_info = { version = "3", default-features = false }
 parking_lot = "0.12.3"
 tokenizers = { path = "../tokenizers" }
 pgrx = "0.12.6"
+reqwest = { version = "0.11.27", features = ["blocking"] }
 rustc-hash = "1.1.0"
 serde = "1.0.210"
 serde_json = "1.0.128"
@@ -41,10 +43,7 @@ tantivy = { git = "https://github.com/paradedb/tantivy.git", package = "tantivy"
 tantivy-common = { git = "https://github.com/paradedb/tantivy.git", rev = "33be46c" }
 thiserror = "1.0.63"
 tracing = "0.1.40"
-utoipa = "4.2.3"
 walkdir = "2.5.0"
-num_cpus = "1.16.0"
-chrono = "0.4.38"
 ordered-float = "4.2.2"
 uuid = "1.10.0"
 strum = { version = "0.26.3" }
