+++ pg_search/src/index/search.rs
@@ -40,7 +40,9 @@ pub const INDEX_TANTIVY_MEMORY_BUDGET: usize = 500_000_000;
 /// PostgreSQL operates in a process-per-client model, meaning every client connection
 /// to PostgreSQL results in a new backend process being spawned on the PostgreSQL server.
 pub static mut SEARCH_EXECUTOR: Lazy<Executor> = Lazy::new(|| {
-    let num_threads = num_cpus::get();
+    let num_threads = std::thread::available_parallelism()
+        .expect("this computer should have at least one CPU")
+        .get();
     Executor::multi_thread(num_threads, "prefix-here").expect("could not create search executor")
 });
