+++ pg_search/src/lib.rs
@@ -25,11 +25,12 @@ mod schema;
 
 #[cfg(test)]
 pub mod fixtures;
+pub mod telemetry;
 
 use self::postgres::customscan;
 use pgrx::*;
 use shared::gucs::PostgresGlobalGucSettings;
-use shared::telemetry::setup_telemetry_background_worker;
+use telemetry::setup_telemetry_background_worker;
 
 // A static variable is required to host grand unified configuration settings.
 pub static GUCS: PostgresGlobalGucSettings = PostgresGlobalGucSettings::new();
@@ -75,7 +76,7 @@ pub unsafe extern "C" fn _PG_init() {
     postgres::options::init();
     GUCS.init("pg_search");
 
-    setup_telemetry_background_worker(shared::telemetry::ParadeExtension::PgSearch);
+    setup_telemetry_background_worker(telemetry::ParadeExtension::PgSearch);
 
     // Register our tracing / logging hook, so that we can ensure that the logger
     // is initialized for all connections.
