+++ pg_search/src/schema/config.rs
@@ -18,7 +18,7 @@
 use crate::postgres::options::SearchIndexCreateOptions;
 use crate::query::SearchQueryInput;
 use pgrx::{JsonB, PgRelation};
-use serde::{de::DeserializeOwned, Deserialize, Deserializer, Serialize};
+use serde::{Deserialize, Serialize};
 use std::str::FromStr;
 
 #[derive(Debug, Serialize, Deserialize, Default, Clone, PartialEq)]
@@ -90,73 +90,6 @@ impl FromStr for SearchConfig {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         let mut deserializer = json5::Deserializer::from_str(s).expect("input is not valid json");
         serde_path_to_error::deserialize(&mut deserializer)
-    }
-}
-// Helpers to deserialize a comma-separated string, following all the rules
-// of csv documents. This let's us easily use syntax like 1,2,3 or one,two,three
-// in the SearchQuery input strings.
-
-#[allow(unused)]
-fn from_csv<'de, D, T>(deserializer: D) -> Result<Vec<T>, D::Error>
-where
-    D: Deserializer<'de>,
-    T: DeserializeOwned + std::str::FromStr,
-    <T as std::str::FromStr>::Err: std::fmt::Debug,
-{
-    deserializer.deserialize_str(CSVVecVisitor::<T>::default())
-}
-
-/// Visits a string value of the form "v1,v2,v3" into a vector of bytes Vec<u8>
-struct CSVVecVisitor<T: DeserializeOwned + std::str::FromStr>(std::marker::PhantomData<T>);
-
-impl<T: DeserializeOwned + std::str::FromStr> Default for CSVVecVisitor<T> {
-    fn default() -> Self {
-        CSVVecVisitor(std::marker::PhantomData)
-    }
-}
-
-impl<'de, T: DeserializeOwned + std::str::FromStr> serde::de::Visitor<'de> for CSVVecVisitor<T>
-where
-    <T as std::str::FromStr>::Err: std::fmt::Debug, // handle the parse error in a generic way
-{
-    type Value = Vec<T>;
-
-    fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(formatter, "a str")
-    }
-
-    fn visit_str<E>(self, s: &str) -> std::result::Result<Self::Value, E>
-    where
-        E: serde::de::Error,
-    {
-        // Treat the comma-separated string as a single record in a CSV.
-        let mut rdr = csv::ReaderBuilder::new()
-            .has_headers(false)
-            .from_reader(s.as_bytes());
-
-        // Try to get the record and collect its values into a vector.
-        let mut output = Vec::new();
-        for result in rdr.records() {
-            match result {
-                Ok(record) => {
-                    for field in record.iter() {
-                        output.push(
-                            field
-                                .parse::<T>()
-                                .map_err(|_| E::custom("Failed to parse field"))?,
-                        );
-                    }
-                }
-                Err(e) => {
-                    return Err(E::custom(format!(
-                        "could not deserialize sequence value: {:?}",
-                        e
-                    )));
-                }
-            }
-        }
-
-        Ok(output)
     }
 }
