+++ pg_search/src/telemetry/controller.rs
