+++ pg_search/src/telemetry/event.rs
