+++ pg_search/src/telemetry/mod.rs
