+++ pg_search/src/telemetry/postgres.rs
