+++ shared/Cargo.toml
@@ -6,40 +6,33 @@ edition = "2021"
 license = "AGPL-3.0"
 
 [features]
-fixtures = ["async-std", "rstest", "soa_derive", "sqlx", "bigdecimal"]
+fixtures = ["async-std", "datafusion", "rstest", "soa_derive", "sqlx", "bigdecimal"]
 
 [dependencies]
+anyhow = "1.0.87"
+bytes = "1.7.1"
+chrono = { version = "0.4.38", features = ["clock", "alloc"] }
+libc = "0.2.159"
 parking_lot = "0.12.3"
 pgrx = "0.12.6"
-reqwest = { version = "0.11.27", features = ["blocking"] }
 serde = "1.0.210"
 serde_json = "1.0.128"
+tempfile = "3.12.0"
 time = { version = "0.3.36", features = ["serde"] }
 tracing = "0.1.40"
+tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
+url = "2.5.2"
 uuid = "1.10.0"
+walkdir = "2.5.0"
+
+# for testing
 async-std = { version = "1.13.0", features = ["attributes"], optional = true }
+bigdecimal = { version = "0.3.1", features = ["serde"], optional = true }
+datafusion = {  version = "38.0.0", optional = true }
 rstest = { version = "0.18.2", optional = true }
 soa_derive = { version = "0.13.0", optional = true }
-sqlx = { version = "0.7.4", features = [
-  "postgres",
-  "runtime-async-std",
-  "time",
-  "bigdecimal",
-  "uuid",
-  "chrono",
-], optional = true }
-bigdecimal = { version = "0.3.1", features = ["serde"], optional = true }
-bytes = "1.7.1"
-url = "2.5.2"
-walkdir = "2.5.0"
-os_info = { version = "3", default-features = false }
-chrono = { version = "0.4.38", features = ["clock", "alloc"] }
-humansize = "2.1.3"
-anyhow = "1.0.87"
-datafusion = "38.0.0"
-tempfile = "3.12.0"
-tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
-libc = "0.2.158"
+sqlx = { version = "0.7.4", features = [ "postgres", "runtime-async-std", "time", "bigdecimal", "uuid", "chrono", ], optional = true }
+
 
 [dev-dependencies]
 mockall = "0.12.1"
